fix discovery legends
